agent: extend reconnect protocol to include a uuid
This uuid is needed if we want to support reconnects from an agent after it
crashed (ie unclean agent shutdown). The uuid is generated on the server side
on the first connect, and saved by the agent to a file. On the next reconnect,
the uuid is presented to the master, who will allow a reconnect even if the
client did not disconnect cleanly before.

Signed-off-by: Olaf Kirch <okir@suse.de>