dbus proxy: fork a subprocess in some configurations
We do this when we service a multiplex downstream socket, and
demux that into separate upstream connections to DBus-daemon.
This is the configuration when running the proxy on the master
host to manage KVM guests, for instance.

This allows us to start only one DBus proxy process for all guests,
rather than having to take care of starting and stopping individual
processes for each guest.

Signed-off-by: Olaf Kirch <okir@suse.de>